+++ vendor_download.sh
@@ -32,6 +32,7 @@ mkdir -p vendor-bins
 
 ASSET_POSTFIX=""
 THEMIS_ASSET_POSTFIX=""
+FICUS_ASSET_POSTFIX=""
 LERNIE_ASSET_POSTFIX=""
 CIRCE_ASSET_POSTFIX=""
 case "$(uname -s)" in
@@ -39,6 +40,7 @@ case "$(uname -s)" in
     case "$(uname -m)" in
       arm64)
         ASSET_POSTFIX="darwin-arm64"
+        FICUS_ASSET_POSTFIX="aarch64-apple-darwin.tar.gz"
         LERNIE_ASSET_POSTFIX="aarch64-macos"
         THEMIS_ASSET_POSTFIX="darwin-arm64"
         CIRCE_ASSET_POSTFIX="aarch64-apple-darwin"
@@ -46,6 +48,7 @@ case "$(uname -s)" in
 
       *)
         ASSET_POSTFIX="darwin-amd64"
+        FICUS_ASSET_POSTFIX="x86_64-apple-darwin.tar.gz"
         LERNIE_ASSET_POSTFIX="x86_64-macos"
         THEMIS_ASSET_POSTFIX="darwin-amd64"
         CIRCE_ASSET_POSTFIX="x86_64-apple-darwin"
@@ -56,24 +59,37 @@ case "$(uname -s)" in
   Linux)
     case "$(uname -m)" in
       aarch64)
-        ASSET_POSTFIX="linux"
-        THEMIS_ASSET_POSTFIX="linux-arm64"
-        LERNIE_ASSET_POSTFIX="aarch64-linux"
-        CIRCE_ASSET_POSTFIX="aarch64-unknown-linux-musl"
-      ;;
-
+        ARCH="aarch64"
+        THEMIS_ARCH="arm64"
+        ;;
+      x86_64)
+        ARCH="x86_64"
+        THEMIS_ARCH="amd64"
+        ;;
       *)
-        ASSET_POSTFIX="linux"
-        THEMIS_ASSET_POSTFIX="linux-amd64"
-        LERNIE_ASSET_POSTFIX="x86_64-linux"
-        CIRCE_ASSET_POSTFIX="x86_64-unknown-linux-musl"
+        echo "Unsupported architecture: $(uname -m)"
+        exit 1
         ;;
     esac
+
+    # Detect libc (musl or gnu)
+    if ldd --version 2>&1 | grep -q musl; then
+      LIBC="musl"
+    else
+      LIBC="gnu"
+    fi
+
+    ASSET_POSTFIX="linux"
+    THEMIS_ASSET_POSTFIX="linux-${THEMIS_ARCH}"
+    FICUS_ASSET_POSTFIX="$ARCH-unknown-linux-$LIBC.tar.gz"
+    LERNIE_ASSET_POSTFIX="${ARCH}-linux"
+    CIRCE_ASSET_POSTFIX="${ARCH}-unknown-linux-musl"
     ;;
   *)
     echo "Warn: Assuming $(uname -s) is Windows"
     ASSET_POSTFIX="windows.exe"
     THEMIS_ASSET_POSTFIX="windows-amd64"
+    FICUS_ASSET_POSTFIX="x86_64-pc-windows-msvc.zip"
     LERNIE_ASSET_POSTFIX="x86_64-windows.exe"
     CIRCE_ASSET_POSTFIX="x86_64-pc-windows-msvc"
     ;;
@@ -141,6 +157,49 @@ done
 echo "Lernie download successful"
 
 rm $LERNIE_RELEASE_JSON
+
+# Download latest release of Ficus
+
+echo "Downloading ficus binary from latest release"
+FICUS_RELEASE_JSON=vendor-bins/ficus-release.json
+curl -sSL \
+    -H "Authorization: token $GITHUB_TOKEN" \
+    -H "Accept: application/vnd.github.v3.raw" \
+    https://api.github.com/repos/fossas/ficus/releases/latest > $FICUS_RELEASE_JSON
+cat "$FICUS_RELEASE_JSON"
+
+FICUS_TAG=$(jq -cr ".name" $FICUS_RELEASE_JSON)
+# Strip the leading 'v' off of the tag
+FICUS_VERSION="${FICUS_TAG/#v/}"
+FILTER=".name == \"ficus-$FICUS_ASSET_POSTFIX\""
+jq -c ".assets | map({url: .url, name: .name}) | map(select($FILTER)) | .[]" $FICUS_RELEASE_JSON | while read -r ASSET; do
+  URL="$(echo "$ASSET" | jq -c -r '.url')"
+  NAME="$(echo "$ASSET" | jq -c -r '.name')"
+  OUTPUT="$(echo vendor-bins/"$NAME" | sed 's/-'"$FICUS_VERSION"'-'$FICUS_ASSET_POSTFIX'$//')"
+
+  echo "Downloading '$NAME' to '$OUTPUT'"
+  curl -sL -H "Authorization: token $GITHUB_TOKEN" -H "Accept: application/octet-stream" -s "$URL" > "$OUTPUT"
+
+  case "$(uname -s)" in
+    Darwin)
+      tar -zxf "$OUTPUT" --strip-components 1 --exclude LICENSE --exclude README.md --directory vendor-bins
+      ;;
+    Linux)
+      tar -zxf "$OUTPUT" --strip-components 1 --exclude LICENSE --exclude README.md --directory vendor-bins
+      ;;
+    *)
+      echo "Warn: Assuming $(uname -s) is Windows"
+      unzip "$OUTPUT" ficus.exe -d vendor-bins
+      mv vendor-bins/ficus.exe vendor-bins/ficus
+      ;;
+  esac
+  rm "$OUTPUT"
+
+
+done
+echo "Ficus download successful"
+
+rm $FICUS_RELEASE_JSON
 
 # Download latest release of Circe